LayoutTests/media/media-controller-playback-expected.txt


RUN(controller = video.controller)
EVENT(canplaythrough)
RUN(controller.play())
EVENT(playing)
EXPECTED (controller.paused == 'false') OK
RUN(controller.currentTime = 5)
EXPECTED (video.currentTime == '5') OK
EXPECTED (video2.currentTime == '5') OK
EVENT(ended)
END OF TEST
